pyproject.toml
[tool.poetry]
name = "boilerplate-fastapi"
version = "0.1.0"
description = ""
authors = ["Bennie ferreira lima"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
databases = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
